Second inspection (from Spanky + Crimson)
Some things were changed here, mainly alot of light placements and firelocks covering maintenance. BUT! There are four main changes I want to cover, which are:

- Robotics has been changed. The Ripley was moved to cargo (along with the clamp) now constructed fully, and some stuff has been moved around
- Medical has a new cryopod! (though it is missing some piping, that spanky has told he will gladly help with (thank you spanky :pray: pls add some cryobeakers too and remove the paper after.....))
- A whole new section of maints south west of the map! Not a lot is added here, but there is the shuttle construction bay, and a brand new substation!
- As moving the shuttle construction bay goes, a new solar panel array was added to where the old one was!

Again, I just wanna reach a hand over to both Spanky and Crimson for helping out reviewing the map. All of the feedback is great and all of the things have been fixed. Enjoy!